separate components into different files
